Fixed visibility issue
If using max-width and max-height properties, and overflow is set to hidden, nested footnotes will not be displayed if they escape their parent note box. If overflow: hidden is removed, the contents of the note box will not be visible, but can still be hovered on, displaying them. Solved by using visibility: hidden and visibility: visible (instead of display: hidden and display: block) with CSS transitions.